feat: add new data structure for portfolio, resume, education, and research information
